+++ src/main/kotlin/thedarkcolour/hardcoredungeons/block/structure/DungeonMobSpawnerBlock.kt
@@ -1,7 +1,0 @@
-package thedarkcolour.hardcoredungeons.block.structure
-
-import net.minecraft.block.Block
-
-class DungeonMobSpawnerBlock(properties: Properties) : Block(properties) {
-
-}