+++ src/main/kotlin/thedarkcolour/hardcoredungeons/registry/RegistryEventHandler.kt
@@ -8,6 +8,7 @@ import net.minecraft.particles.ParticleType
 import net.minecraft.potion.Effect
 import net.minecraft.util.SoundEvent
 import net.minecraft.world.biome.Biome
+import net.minecraft.world.gen.feature.Feature
 import net.minecraft.world.gen.feature.structure.Structure
 import net.minecraft.world.gen.surfacebuilders.SurfaceBuilder
 import net.minecraftforge.event.RegistryEvent.Register
@@ -36,6 +37,7 @@ object RegistryEventHandler {
         MOD_BUS.addGenericListener(::registerDataSerializers)
         MOD_BUS.addGenericListener(::registerEffects)
         MOD_BUS.addGenericListener(::registerEntities)
+        MOD_BUS.addGenericListener(::registerFeatures)
         MOD_BUS.addGenericListener(::registerStructures)
         MOD_BUS.addGenericListener(::registerItems)
         MOD_BUS.addGenericListener(::registerParticles)
@@ -59,6 +61,8 @@ object RegistryEventHandler {
     private fun registerEffects(event: Register<Effect>) = HEffects.registerEffects(event.registry)
 
     private fun registerEntities(event: Register<EntityType<*>>) = HEntities.registerEntities(event.registry)
+
+    private fun registerFeatures(event: Register<Feature<*>>) = HFeatures.registerFeatures(event.registry)
 
     private fun registerStructures(event: Register<Structure<*>>) = HStructures.registerStructures(event.registry)