+++ src/main/kotlin/thedarkcolour/hardcoredungeons/tileentity/DungeonSpawnerTileEntity.kt
@@ -1,0 +1,11 @@
+package thedarkcolour.hardcoredungeons.tileentity
+
+import net.minecraft.tileentity.TileEntity
+import net.minecraft.tileentity.TileEntityType
+import thedarkcolour.hardcoredungeons.registry.HTileEntities
+
+class DungeonSpawnerTileEntity(tileEntityTypeIn: TileEntityType<*> = HTileEntities.DUNGEON_SPAWNER) : TileEntity(tileEntityTypeIn) {
+    val remainingKills = 0
+
+
+}